EWM-458. Json file config
