+++ lib/app/service/bootstrap/bootstrap_service.dart
@@ -1,5 +1,5 @@
 import 'package:app/app/service/bootstrap/bootstrap_steps.dart';
-import 'package:app/app/service/connection/presets_connection_service.dart';
+import 'package:app/app/service/presets_connection/presets_connection_service.dart';
 import 'package:app/bootstrap/bootstrap.dart';
 import 'package:app/core/app_build_type.dart';
 import 'package:app/event_bus/events/bootstrap/bootstrap_event.dart';
