+++ lib/feature/network/bloc/manage_networks/manage_networks_bloc.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:app/app/service/connection/data/connection_data/connection_data.dart';
-import 'package:app/app/service/connection/presets_connection_service.dart';
+import 'package:app/app/service/presets_connection/presets_connection_service.dart';
 import 'package:app/app/service/service.dart';
 import 'package:app/core/bloc/bloc_mixin.dart';
 import 'package:bloc/bloc.dart';
